Rename ErrorBoundary component
