+++ src/App/bootstrap/app-stage2.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Route, Switch } from "react-router-dom"
 import AccountPage from "~Account/components/AccountView"
 import SettingsPage from "~AppSettings/components/AppSettingsView"
-import ErrorBoundary from "~Generic/components/ErrorBoundary"
+import ErrorBoundary from "~Generic/components/ErrorBoundaryView"
 import { VerticalLayout } from "~Layout/components/Box"
 import { appIsLoaded } from "~SplashScreen/splash-screen"
 import ConnectionErrorListener from "~Toasts/components/ConnectionErrorListener"
